came back to a context
realise that I will have to pass functions and other things from the coffeeCard universe to other components and it will be simpler with a context so I cameBack to a context and put all that was in the shared folder in the context, i'm still not realy using the as a context but now I will star to increment all the other things needed in this context and it will be better.
also: requre is still not working so my solution is continue to import all the images in the context.
I still get a error in the console that I will adress next.