fixed infinity loop bug
I had an infinity loop bug using the useEffect to update the counterStates, but the useEffect was disposable in this case since i'm updating the counterStates directly in the function of the onClick event. since I can pass the state already updated to the coffeeCard and from it to the coffeeList it was the best option to get rid of the useEffect and fix the bug this way.