moved coffecardstates to context
moved the states of coffeeCard to the context this way the context will contain all that is exclusive of the coffeeCard universe.
started to use the useContext to acess the informations of the context and also the createContext in the coffeeCardContext.
will move other functions later? I chose to move one by one to keep track of the app and if something brokes be able to be more precise on what caused the error